authenticate/requirements.txt
cryptography
pyotp
pydantic
psycopg2-binary
qrcode
Pillow